Split out the TriviaBot class to its own file.
